+++ 30_sources/SRC_Zupfnoter/src/user-interface.js
@@ -391,8 +391,8 @@ function init_w2ui(uicontroller) {
     name: 'toolbar',
     style: tbstyle,
     items: [
-      {type: 'button', id: 'tb_home', icon: 'fa fa-home', text: 'Zupfnoter'},
-      {type: 'html', html: '<div style="width:25px"/>'},
+      {type: 'button', id: 'tb_home', icon: 'fa fa-info-circle', text: ''},
+     // {type: 'html', html: '<div style="width:25px"/>'},
       {
         type: 'menu',
         id: 'tb_file',
@@ -897,14 +897,10 @@ function init_w2ui(uicontroller) {
         id: 'sb_mode',
         text: '<div style="padding: 0px !important;"><span class="sb-mode"></span></div>'
       },
-      {
-        type: 'button',
-        id: 'sb_dropbox-status',
-        text: '<div style="padding: 0px !important;"><span class="dropbox-status"></span></div>'
-      },
+
       {
         type: 'menu',
-        id: 'sb_setdropboxpath',
+        id: 'sb_dropbox_status',
         tooltip: "choose drobox path",
         onClick: function (event) {
           if (event.subItem) {
@@ -1226,8 +1222,8 @@ function update_settings_menu(settings) {
 
 
 function update_systemstatus_w2ui(systemstatus) {
-  $(".dropbox-status").html(systemstatus.dropbox);
-  w2ui.layout_top_toolbar.set('tb_home', {text: "Zupfnoter" + " " + systemstatus.version})
+  w2ui.layout_statusbar_main_toolbar.set("sb_dropbox_status", {text: systemstatus.dropbox})
+  w2ui.layout_top_toolbar.set('tb_home', {tooltip: "Zupfnoter" + " " + systemstatus.version})
 
   var tb_view_title = w2ui.layout_top_toolbar.get('tb_view:' + systemstatus.view)
   tb_view_title = (tb_view_title ? tb_view_title.text : systemstatus.view)
@@ -1271,7 +1267,7 @@ function update_systemstatus_w2ui(systemstatus) {
     return {text: i}
   });
 
-  w2ui.layout_statusbar_main_toolbar.get("sb_setdropboxpath").items = dropbox_path_menu_items;
+  w2ui.layout_statusbar_main_toolbar.get("sb_dropbox_status").items = dropbox_path_menu_items;
 
 }